Refactored Services to be Use Cases, no Data Access code. Split Service interface into three interfaces and three implementations, Created Three repository interfaces and JPA implementations. Created Named Queries and put them into the Domain objects. Moved all styling of web into styles.css file and moved all JavaScript from .xhtml to a .js file. Also removed snapshot repository from the pom file. Having the Milestone and Snapshot Spring repositories in the pom.xml causes a crap load of .pom for all snapshots and milestones of Spring modules. Even though the pom isn't using any of them. It doesn't download jars, but just pom.xml

+++ server/api/src/main/java/com/springsource/html5expense/services/Flag.java
@@ -17,16 +17,16 @@ package com.springsource.html5expense.services;
 
 public class Flag {
 
-    private final Integer expenseId;
+    private final Long expenseId;
 
     private final String value;
 
-    public Flag(Integer expenseId, String value) {
+    public Flag(Long expenseId, String value) {
         this.expenseId = expenseId;
         this.value = value;
     }
 
-    public Integer getExpenseId() {
+    public Long getExpenseId() {
         return expenseId;
     }
